requirements.txt
# absl-py==2.2.1
# altair==5.5.0
# annotated-types==0.7.0
# asttokens==3.0.0
# astunparse==1.6.3
# attrs==25.3.0
# blinker==1.9.0
# cachetools==5.5.2
# certifi==2025.1.31
# charset-normalizer==3.4.1
# click==8.1.8
# comm==0.2.2
# contourpy==1.3.1
# cycler==0.12.1
# debugpy==1.8.13
# decorator==5.2.1
# executing==2.2.0
# Flask==3.1.0
# flask-cors==5.0.1
# flatbuffers==25.2.10
# fonttools==4.57.0
# gast==0.6.0
# gitdb==4.0.12
# GitPython==3.1.44
# google-ai-generativelanguage==0.6.15
# google-api-core==2.24.2
# google-api-python-client==2.166.0
# google-auth==2.38.0
# google-auth-httplib2==0.2.0
# google-generativeai==0.8.4
# google-pasta==0.2.0
# googleapis-common-protos==1.69.2
# grpcio==1.71.0
# grpcio-status==1.71.0
# h5py==3.13.0
# httplib2==0.22.0
# idna==3.10
# ipykernel==6.29.5
# ipython==9.1.0
# ipython_pygments_lexers==1.1.1
# itsdangerous==2.2.0
# jedi==0.19.2
# Jinja2==3.1.6
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# keras==3.9.2
# kiwisolver==1.4.8
# libclang==18.1.1
# Markdown==3.7
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# matplotlib==3.10.1
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# ml_dtypes==0.5.1
# namex==0.0.8
# narwhals==1.33.0
# nest-asyncio==1.6.0
# numpy==2.1.3
# opt_einsum==3.4.0
# optree==0.14.1
# packaging==24.2
# pandas==2.2.3
# parso==0.8.4
# pexpect==4.9.0
# pillow==11.1.0
# platformdirs==4.3.7
# prompt_toolkit==3.0.50
# proto-plus==1.26.1
# protobuf==5.29.4
# psutil==7.0.0
# ptyprocess==0.7.0
# pure_eval==0.2.3
# pyarrow==19.0.1
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pydantic==2.11.1
# pydantic_core==2.33.0
# pydeck==0.9.1
# Pygments==2.19.1
# pyparsing==3.2.3
# python-dateutil==2.9.0.post0
# pytz==2025.2
# pyzmq==26.4.0
# referencing==0.36.2
# requests==2.32.3
# rich==14.0.0
# rpds-py==0.24.0
# rsa==4.9
# setuptools==78.1.0
# six==1.17.0
# smmap==5.0.2
# stack-data==0.6.3
# streamlit==1.44.1
# tenacity==9.1.2
# tensorboard==2.19.0
# tensorboard-data-server==0.7.2
tensorflow==2.19.0
# termcolor==3.0.1
# toml==0.10.2
# tornado==6.4.2
# tqdm==4.67.1
# traitlets==5.14.3
# typing-inspection==0.4.0
# typing_extensions==4.13.0
# tzdata==2025.2
# uritemplate==4.1.1
# urllib3==2.3.0
# watchdog==6.0.0
# wcwidth==0.2.13
# Werkzeug==3.1.3
# wheel==0.45.1
# wrapt==1.17.2
flask
streamlit
pandas
numpy
scikit-learn
joblib
Pillow
